change handling of assignment statements and sln
The way assignment statements are handled in the Interpreter now lets
us save values to arrays, but the implementation is still not quite
right.

Also, in the Visual Studio Build Configuration Manager, in the
ShankCompiler project context, I changed the configuration from
"Release" to "Debug." This allowed variable values to update dynamically
during debugging, but I think it also made the "dotnet run" command
think a carriage return is being passed in as a command line argument.
If you want to run the project with "dotnet run" now, make sure you
include a directory or file as a command line argument, because
running the command without any arguments no longer works.