null-checking; end-of-line logic
Add GetIdentifierValue method to Token to consolidate null-checking.

Add ExpectsEndOfLine, RequiresEndOfLine, and ConsumeBlankLines methods to Parser.

Move "consume blank lines" logic into MatchAndRemove as a temporary fix for allowing blank lines anywhere. We don't want MatchAndRemove to be the permanent home of this logic because we want to keep MAR simple. The eventual permanent fix will be to convert every "MatchAndRemove(EndOfLine)" call to a "ExpectsEndOfLine()" or a "RequiresEndOfLine()" call.