OutputHelper; isRecord flag; shorten Statement()
Create OutputHelper class to write debug output to a file.

Try to parse a statement as a record member declaration based on the value of an "isRecord" flag. We need to use this flag because two arbitrary identifiers in a row could be a record member declaration or a function call from the Parser's point of view, so we need to know the context.

Make Parser.Statement() more concise.