Lexer refactors; Invalid cast fix
A few more small Lexer refactors.

Quick fix for an issue in SemanticAnalysis.CheckComparisons where
InvalidCastException was being thrown if the right side of a boolean
expression wasn't a VariableReferenceNode. We should make a better fix
for this in the future.

Add revision of Shank Language Definition document to repository.

Create new example Shank program for testing negatives and positives and
boolean expressions.