move IR to APPDATA; ILlvmTranslatable
Introduce ILlvmTranslatable interface to avoid code duplication.

Running any Shank file now creates the file "$($env:APPDATA)/IrOutput.ll".
Currently, this ll file's contents will be the same for every Shank file EXCEPT
Shank files that contain a function named "validForLlvm".

If the Shank file does not contain such a function, running
PS> lli "$($env:APPDATA)/IrOutput.ll"
should print "hello invalid" to the console. (PS> is the Powershell prompt.)

If the Shank file does contain such a function, running the same command as
above will generate IR based on the contents of the Shank file. So far I have
only given it the ability to generate IR for calls to Shank's Builtin `write'
passing in a string. But you should be able to pass any arbitrary string to
`write' (as long as you call `write' in `start') and if you run the resulting
ll file with `lli' it should print that string to the console.