LLVM printing ints; testing strings
Experiment with adding functions with duplicate names to an LLVM module.

Create IntNode case in IrGeneratorByNode that just converts the IntNode to an
LLVM Global String Ptr. This is to make something like `write 2' work in Shank,
but it's not a good approach overall because we want to make actual LLVm
integers from IntNodes in other situations.

Add an example Shank program for passing a string to a function to the
Powershell script.