move enum type definitions out of ASTNode
Refactor-move definitions of enum type VrnExtType, BooleanExpressionOpType, and
MathOpType from ASTNode back to their more logical homes (VariableReferenceNode,
BooleanExpressionNode, and MathOpNode respectively).

I moved these definitions to ASTNode when I first added source generation to the
repo because I was getting errors, but now that I've moved them back, the errors
have not reappeared, even though source generation is still in place. I'm not
sure why this is, but it could be related to the fact that we refactored all the
node classes into their own files in the meantime.