VDN constructors; VDN TypeName; real ranges
Create parameterless constructor in VariableDeclarationNode so we can start
creating parametered constructors and not break the object initializers.

Add TypeName property to VariableDeclarationNode so we can see the type name in
the AST JSON.

Add minimal program for testing real type ranges to dotShank/RealRanges and the
Powershell script. When you declare a variable of type `real' and give it a
range with constant numbers with decimal points, it (I think incorrectly) throws
a SyntaxErrorException saying it expects an integer range.