VunVsTypeCheckingVisitor and other stuff
Create VunVsTypeCheckingVisitor and test it in SA.NewCheckAssignment.

Add one param-validating method to VariableUsageIndexNode for enforcing
invariants (planning to add more).

Add expression-bodied property to VariableUsageIndexNode to get the "base name"
into which the index indexes.

Add "GetBaseName" method to VariableUsageMemberNode to accomplish something
similar, but it is not correct yet.