EnumNode; ExpressionTypingVisitor
Make EnumNode constructor create VDNs for the members and have
VariablesGettingVisitor visit enums and put those VDNs in the VariablesInScope
property on FunctionNode.

Make ExpressionTypingVisitor that evaluates the type of every expression in the
AST and saves it in a property on ExpressionNode called Type.

Rename EnumNode.Type to EnumNode.EType so it doesn't shadow ExpressionNode.Type.