DefaultType; refersTo typing issue; UnknownTypesVisitor issue
Make DefaultType Type so the properties that don't hold a Type until after
they're visited don't need to be nullable (let me know if this causes any
issues).

Diagnose issue with getting the type of a vun that is a reference.

Experiment with running just the UnknownTypesVisitor before the
ExpressionTypingVisitor, but the former does not seem to resolve all the
unknowns that old SA does when there are references.