VUN.GetMyType; vuop stuff close to done
Use adapted version of Mendel's version of VUN.GetInnerType and now all the
args lines in ShankTestPaths_main_vuop pass SA (thanks Mendel).

Only thing that is left is to fix where the nested records program with
references fails in interpreter and the vuop stuff should be pretty much done.