updating parameters in Interpreter
Add comments explaining the logic of the if-statement in the loop in Interpreter
that updates the passed-in parameters after a function call.

Handle updating these parameters with the vuop property moves.

next: figure out a way to update only part of an IDT if a passed-in param is
only part of a VUN.